split reducers into two files
