+++ frontend/src/components/home/DisplayTagArea.js
@@ -10,16 +10,16 @@ class DisplayTagArea extends React.Component {
     }
 
     createTags() {
-        let ret  = [];
+        let ret = [];
         let arr = this.props.tags.slice();
         let len = arr.length;
-        for (let i = 0; i < len; i++)  {
-            let details = {tag: arr[i].tag, selected: arr[i].selected, color: arr[i].color};
-            let selectedStatus = arr[i].selected?"selectedTag":"";
-            ret.push(<span style={arr[i].selected?{background: arr[i].color}: {}}
-                           className={"tagItem " + selectedStatus} key={i}>
-                <Button className = "Button x details" name = "x" funcName = "deleteTag" tagIndex = {i} />
-                <Tag tagDetails = {{...details}} tagIndex = {i}
+        for (let i = 0; i < len; i++) {
+            let details = { tag: arr[i].tag, selected: arr[i].selected, color: arr[i].color };
+            let selectedStatus = arr[i].selected ? "selectedTag" : "";
+            ret.push(<span style={arr[i].selected ? { background: arr[i].color } : {}}
+                className={"tagItem " + selectedStatus} key={i}>
+                <Button className="Button x details" name="x" funcName="deleteTag" tagIndex={i} />
+                <Tag tagDetails={{ ...details }} tagIndex={i}
                 /></span>);
         }
         return ret;
@@ -36,7 +36,7 @@ class DisplayTagArea extends React.Component {
 }
 
 const mapStateToProps = (state) => {
-    return {tags: state.tags};
+    return { tags: state.tags };
 };
 
 export default connect(mapStateToProps)(DisplayTagArea);
