+++ frontend/src/components/home/reducers/InputReducer.js
@@ -1,0 +1,8 @@
+const inputReducer =  (init = "", action) => {
+    if (action.type === "INPUT_TAG") {
+        return action.input;
+    }
+    return init;
+ };
+
+  export default inputReducer;
